+++ tests/engine/pieces/pawn.spec.ts
@@ -1,15 +1,16 @@
 import 'chai/register-should';
 import Pawn from '../../../src/engine/pieces/pawn';
+import Rook from '../../../src/engine/pieces/rook';
 import Board from '../../../src/engine/board';
 import Player from '../../../src/engine/player';
 import Square from '../../../src/engine/square';
 
 describe('Pawn', () => {
 
-    describe('white pawns', () => {
+    let board: Board;
+    beforeEach(() => board = new Board());
 
-        let board: Board;
-        beforeEach(() => board = new Board());
+    describe('white pawns', () => {
 
         it('can only move one square up if they have already moved', () => {
             const pawn = new Pawn(Player.WHITE);
@@ -60,6 +61,28 @@ describe('Pawn', () => {
             moves.should.deep.include.members([Square.at(4, 7), Square.at(5, 7)]);
         });
 
+    });
+
+    it('cannot move if there is a piece in front', () => {
+        const pawn = new Pawn(Player.BLACK);
+        const blockingPiece = new Rook(Player.WHITE);
+        board.setPiece(Square.at(6, 3), pawn);
+        board.setPiece(Square.at(5, 3), blockingPiece);
+
+        const moves = pawn.getAvailableMoves(board);
+
+        moves.should.be.empty;
+    });
+
+    it('cannot move two squares if there is a piece two sqaures in front', () => {
+        const pawn = new Pawn(Player.BLACK);
+        const blockingPiece = new Rook(Player.WHITE);
+        board.setPiece(Square.at(6, 3), pawn);
+        board.setPiece(Square.at(4, 3), blockingPiece);
+
+        const moves = pawn.getAvailableMoves(board);
+
+        moves.should.not.deep.include(Square.at(4, 3));
     });
 
 });
